backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
brotli==1.2.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cryptography==46.0.5
cssselect2==0.9.0
distro==1.9.0
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.129.0
fonttools==4.61.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.13.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==1.2.10
langchain-core==1.2.16
langgraph==1.0.10
langgraph-checkpoint==4.0.1
langgraph-prebuilt==1.0.8
langgraph-sdk==0.3.9
langsmith==0.7.9
openai==2.21.0
orjson==3.11.7
ormsgpack==1.12.2
packaging==26.0
pdfminer.six==20251230
pdfplumber==0.11.9
pillow==12.1.1
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pydyf==0.12.1
pypdfium2==5.4.0
pyphen==0.17.2
python-dotenv==1.2.1
python-multipart==0.0.22
PyYAML==6.0.3
requests==2.32.5
requests-toolbelt==1.0.0
resend==2.23.0
sniffio==1.3.1
SQLAlchemy==2.0.47
sqlmodel==0.0.37
starlette==0.52.1
tenacity==9.1.4
tinycss2==1.5.1
tinyhtml5==2.0.0
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uuid_utils==0.14.1
uvicorn==0.40.0
weasyprint==68.1
webencodings==0.5.1
xxhash==3.6.0
zopfli==0.4.1
zstandard==0.25.0
